Fixes #1649: Map.content overrides
